[7.0] Fix connector redmine
In vanilla Odoo, the timesheet is not clever. Whenever the user adds a new entry or changes a value in the widget, all previous lines are unlinked and new lines are created for all previous existing lines

Remove read access to the Redmine connector for standard users

Fix method to import timesheets

An employee should be able to import his own timesheets but not have access to the connector

For that reason, we call SUPERUSER_ID but we check that the employee has write access to the requested timesheets.

Change default timesheet line name imported from redmine

Fix description of timesheets

Add explicit mapping of field general_account_id

Fix error message

Prevent module connector from breaking tests in connector Redmine

Add mapping of product, product_uom and amount