[WIP] Interpret a no-billing flag set in Redmine
Redmine permits custom fields; we use one such custom field called
'Admin flags' to indicate if time spent should be billable.  This same
concept is available in Odoo with the 'to_invoice' property of the an
analytic line. As it is painful to manually fix-up time imported from
Redmine for this property only, also preserve this status during
importation.

---
This is untested (sorry! What is the easiest way to test this?) and
relies on some hard-coded values (u'Admin flags' and u'2') relevant in
my particular situation. Suggestions for improvement most welcome!