Prevent inter-company importation
If the user and the analytic account are linked to different companies,
don't import this time entry. The 'code'/'reference' field of an an
analytic account is unique even among multiple companies. Depending on
the value matched from Redmine(s), time may be attributed to analytic
accounts in a different company.

The goal here is to solve a particular problem where users in a related
company (using the same Redmine in this case but regardless), are
importing their time to accounts belong to a parent company but having
no visibility that this has occured.
---
This is untested (sorry! What is the easiest way to test this?) and,
rather than throwing a MappingError, it would be niceer to simply drop
the offending record.
Suggestions for improvement most welcome!